Move data transformation into LocationTable
